# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/developer/api-keys": {
      "get": {
        "description": "Lists every key owned by the calling integrator. Secrets are never returned — only fingerprints + metadata.",
        "operationId": "ApiKeysController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyRecordDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid operator JWT."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List API keys",
        "tags": [
          "developer"
        ]
      },
      "post": {
        "description": "Issues a new bearer secret. The `secret` field is returned EXACTLY ONCE; the server only stores its HMAC fingerprint. Use the secret as `Authorization: Bearer <secret>` on subsequent calls (FR-313).",
        "operationId": "ApiKeysController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreatedOnceDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid operator JWT."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an API key",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/developer/api-keys/{id}/rotate": {
      "post": {
        "description": "Atomically revokes the old key and issues a new secret with the same name, scopes, and environment. Returns both the revoked metadata and the newly created key (secret shown once).",
        "operationId": "ApiKeysController_rotate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotateResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid operator JWT."
          },
          "404": {
            "description": "Unknown key id, foreign integrator, or already revoked."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rotate an API key",
        "tags": [
          "developer"
        ]
      }
    },
    "/api/v1/developer/api-keys/{id}": {
      "delete": {
        "description": "Immediately revokes the key. Any in-flight request with the old secret will be rejected on its next call.",
        "operationId": "ApiKeysController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRecordDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid operator JWT."
          },
          "404": {
            "description": "Unknown key id or foreign integrator."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an API key",
        "tags": [
          "developer"
        ]
      }
    },
...
    "/api/v1/custom-rules": {
      "get": {
        "description": "Returns summaries (no match/card body). Each rule fires at evaluation time when the match block holds.",
        "operationId": "CustomRulesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List custom CDS rules for the current integrator",
        "tags": [
          "custom-rules"
        ]
      },
      "post": {
        "description": "Define a match block (medications / diagnoses / allergies / labs / patient signals) and a card. The agentic engine merges the resulting cards with the signed bundle + policies.",
        "operationId": "CustomRulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a custom rule",
        "tags": [
          "custom-rules"
        ]
      }
    },
    "/api/v1/custom-rules/{id}": {
      "get": {
        "operationId": "CustomRulesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single rule with its full match + card body",
        "tags": [
          "custom-rules"
        ]
      },
      "patch": {
        "operationId": "CustomRulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a custom rule (partial)",
        "tags": [
          "custom-rules"
        ]
      },
      "delete": {
        "operationId": "CustomRulesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a custom rule by id",
        "tags": [
          "custom-rules"
        ]
      }
    },
...
  },
  "info": {
    "title": "VedaMD API",
    "description": "Source-available, stateless Clinical Decision Support API. CDS Hooks + REST/JSON. VedaMD never stores patient data — records remain with your EMR.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiKeyRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4e5f60718"
          },
          "integratorId": {
            "type": "string",
            "example": "integrator_acme"
          },
          "name": {
            "type": "string",
            "example": "EMR production — Nairobi cluster"
          },
          "fingerprint": {
            "type": "string",
            "description": "HMAC-SHA-256 fingerprint of the secret. Used for safe display.",
            "example": "7e3d…"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "cds:evaluate",
                "cds:discover",
                "content:read",
                "drug-info:read",
                "terminology:read",
                "bundles:read",
                "integration-log:read",
                "clinical-audit:run"
              ]
            }
          },
          "environment": {
            "type": "string",
            "enum": [
              "sandbox",
              "production"
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-25T10:00:00.000Z"
          },
          "lastUsedAt": {
            "type": "string",
            "example": "2026-05-25T11:42:00.000Z"
          },
          "revokedAt": {
            "type": "string",
            "example": "2026-05-25T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "integratorId",
          "name",
          "fingerprint",
          "scopes",
          "environment",
          "createdAt"
        ]
      },
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable label for this key — shown in the dashboard.",
            "example": "EMR production — Nairobi cluster",
            "minLength": 1,
            "maxLength": 80
          },
          "scopes": {
            "type": "array",
            "description": "Capability scopes granted to this key. Must contain at least one.",
            "example": [
              "cds:evaluate",
              "content:read"
            ],
            "items": {
              "type": "string",
              "enum": [
                "cds:evaluate",
                "cds:discover",
                "content:read",
                "drug-info:read",
                "terminology:read",
                "bundles:read",
                "integration-log:read",
                "clinical-audit:run"
              ]
            }
          },
          "environment": {
            "type": "string",
            "description": "Which environment this key applies to. Sandbox keys are isolated from prod.",
            "enum": [
              "sandbox",
              "production"
            ],
            "example": "sandbox"
          }
        },
        "required": [
          "name",
          "scopes",
          "environment"
        ]
      },
      "ApiKeyCreatedOnceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "a1b2c3d4e5f60718"
          },
          "integratorId": {
            "type": "string",
            "example": "integrator_acme"
          },
          "name": {
            "type": "string",
            "example": "EMR production — Nairobi cluster"
          },
          "fingerprint": {
            "type": "string",
            "description": "HMAC-SHA-256 fingerprint of the secret. Used for safe display.",
            "example": "7e3d…"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "cds:evaluate",
                "cds:discover",
                "content:read",
                "drug-info:read",
                "terminology:read",
                "bundles:read",
                "integration-log:read",
                "clinical-audit:run"
              ]
            }
          },
          "environment": {
            "type": "string",
            "enum": [
              "sandbox",
              "production"
            ]
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-25T10:00:00.000Z"
          },
          "lastUsedAt": {
            "type": "string",
            "example": "2026-05-25T11:42:00.000Z"
          },
          "revokedAt": {
            "type": "string",
            "example": "2026-05-25T12:00:00.000Z"
          },
          "secret": {
            "type": "string",
            "description": "The raw bearer secret. Shown EXACTLY ONCE — copy it now. Use it as `Authorization: Bearer <secret>`.",
            "example": "vmd_test_8c0w7Z…"
          }
        },
        "required": [
          "id",
          "integratorId",
          "name",
          "fingerprint",
          "scopes",
          "environment",
          "createdAt",
          "secret"
        ]
      },
      "RotateResponseDto": {
        "type": "object",
        "properties": {
          "revoked": {
            "$ref": "#/components/schemas/ApiKeyRecordDto"
          },
          "created": {
            "$ref": "#/components/schemas/ApiKeyCreatedOnceDto"
          }
        },
        "required": [
          "revoked",
          "created"
        ]
      },
...
      "CustomRuleMatchDto": {
        "type": "object",
        "properties": {
          "medications": {
            "example": [
              "metformin"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "diagnoses": {
            "example": [
              "ckd",
              "chronic kidney"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hooks": {
            "example": [
              "medication-prescribe"
            ],
            "description": "Restrict the rule to specific CDS hooks. Empty = all hooks.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ageMinYears": {
            "type": "number",
            "example": 0
          },
          "ageMaxYears": {
            "type": "number",
            "example": 18
          },
          "pregnant": {
            "type": "boolean"
          },
          "eGFRBelow": {
            "type": "number",
            "example": 30,
            "description": "Rule matches if eGFR < this threshold."
          },
          "labs": {
            "description": "Array of lab-value clauses (all must hold).",
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        }
      },
      "CustomRuleCardDto": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string",
            "example": "Hold metformin if eGFR < 30 — local renal protocol.",
            "minLength": 4,
            "maxLength": 240
          },
          "detail": {
            "type": "string",
            "description": "Markdown-rendered detail body."
          },
          "sourceLabel": {
            "type": "string",
            "example": "Acme Hospital Pharmacy SOP v3.1"
          },
          "sourceUrl": {
            "type": "string",
            "example": "https://intranet.acme/sops/renal"
          },
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "links": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "summary"
        ]
      },
      "CreateCustomRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Hold metformin in CKD 4/5",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "indicator": {
            "type": "string",
            "enum": [
              "critical",
              "warning",
              "info"
            ]
          },
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "match": {
            "$ref": "#/components/schemas/CustomRuleMatchDto"
          },
          "card": {
            "$ref": "#/components/schemas/CustomRuleCardDto"
          }
        },
        "required": [
          "name",
          "indicator",
          "match",
          "card"
        ]
      },
      "UpdateCustomRuleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "indicator": {
            "type": "string",
            "enum": [
              "critical",
              "warning",
              "info"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "match": {
            "$ref": "#/components/schemas/CustomRuleMatchDto"
          },
          "card": {
            "$ref": "#/components/schemas/CustomRuleCardDto"
          }
        }
      },
...
    }
  }
}